+++ src/udp.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashSet, net::IpAddr, sync::Arc};
+use std::{collections::HashSet, mem::size_of, net::IpAddr, sync::Arc};
 
 use crate::{
     connection_handler::ConnectionHandler,
@@ -21,13 +21,13 @@ use tokio::{
     select,
 };
 
-pub const MAX_PACKET_SIZE: usize = std::mem::size_of::<u16>() + u16::MAX as usize;
+pub const MAX_PACKET_SIZE: usize = size_of::<u16>() + u16::MAX as usize;
 
-/// Creates and returns a buffer on the heap with enough space to contain any possible
-/// UDP datagram.
-///
-/// This is put on the heap and in a separate function to avoid the 64k buffer from ending
-/// up on the stack and blowing up the size of the futures using it.
+// Creates and returns a buffer on the heap with enough space to contain any possible
+// UDP datagram.
+//
+// This is put on the heap and in a separate function to avoid the 64k buffer from ending
+// up on the stack and blowing up the size of the futures using it.
 #[inline]
 fn datagram_buffer() -> Box<[u8; MAX_PACKET_SIZE]> {
     Box::new([0u8; MAX_PACKET_SIZE])
@@ -74,8 +74,8 @@ impl UdpPortHandler for Arc<UdpHandler> {
             let mut socket = socket;
             loop {
                 let mut buf = datagram_buffer();
-                match socket.peek_from(buf.as_mut()).await {
-                    Ok((_, address)) => {
+                match socket.recv_from(buf.as_mut()).await {
+                    Ok((len, address)) => {
                         let ip = address.ip().to_canonical();
                         if !clone.ip_filter.is_allowed(ip) {
                             #[cfg(not(coverage_nightly))]
@@ -100,34 +100,51 @@ impl UdpPortHandler for Arc<UdpHandler> {
                                     .into_bytes(),
                                 );
                             }
+                            let socket = std::mem::replace(
+                                &mut socket,
+                                get_udp_socket((listen_address, port))
+                                    .expect("should re-create UDP socket"),
+                            );
                             if let Err(error) = socket.connect(address).await {
                                 #[cfg(not(coverage_nightly))]
                                 tracing::error!(%port, %error, "Error connecting UDP socket.");
                                 continue;
                             }
-                            let connected_socket = socket;
-                            socket = get_udp_socket((listen_address, port)).expect("");
+                            let mut read_buf = datagram_buffer();
+
+                            // Prepare already consumed data to be sent to SSH channel
+                            *read_buf[..size_of::<u16>()]
+                                .as_mut_array()
+                                .expect("length checked") = (len as u16).to_be_bytes();
+                            read_buf[size_of::<u16>()..size_of::<u16>() + len]
+                                .copy_from_slice(&buf[..len]);
+
                             tokio::spawn(async move {
-                                let udp_read = Arc::new(connected_socket);
+                                let udp_read = Arc::new(socket);
                                 let udp_write = udp_read.clone();
                                 let (mut ssh_read, mut ssh_write) = tokio::io::split(channel);
 
                                 let udp2ssh = async move {
-                                    let mut buf = datagram_buffer();
+                                    let mut buf = read_buf;
+                                    if let Err(error) =
+                                        ssh_write.write_all(&buf[..size_of::<u16>() + len]).await
+                                    {
+                                        #[cfg(not(coverage_nightly))]
+                                        tracing::warn!(%port, %error, "Error writing to SSH channel for UDP.");
+                                        return;
+                                    }
                                     loop {
                                         match udp_read
-                                            .recv(&mut buf.as_mut()[std::mem::size_of::<u16>()..])
+                                            .recv(&mut buf.as_mut()[size_of::<u16>()..])
                                             .await
                                         {
                                             Ok(len) => {
-                                                *buf[..std::mem::size_of::<u16>()]
+                                                *buf[..size_of::<u16>()]
                                                     .as_mut_array()
                                                     .expect("length checked") =
                                                     (len as u16).to_be_bytes();
                                                 if let Err(error) = ssh_write
-                                                    .write_all(
-                                                        &buf[..std::mem::size_of::<u16>() + len],
-                                                    )
+                                                    .write_all(&buf[..size_of::<u16>() + len])
                                                     .await
                                                 {
                                                     #[cfg(not(coverage_nightly))]
@@ -184,7 +201,7 @@ impl UdpPortHandler for Arc<UdpHandler> {
                     }
                     Err(error) => {
                         #[cfg(not(coverage_nightly))]
-                        tracing::warn!(%port, %error, "Error reading from UDP port.")
+                        tracing::warn!(%port, %error, "Error getting remote connection on UDP port.")
                     }
                 }
             }