+++ tests/integration/udp_allow_requested_ports.rs
@@ -156,13 +156,10 @@ async fn udp_allow_requested_ports() {
         .await
         .expect("Local forwarding failed");
     if timeout(Duration::from_secs(5), async {
-        channel
-            .data(&[0x00, 0x04, b'P', b'i', b'n', b'g'][..])
-            .await
-            .unwrap();
+        channel.data(&b"\x00\x04Ping"[..]).await.unwrap();
         match &mut channel.wait().await.unwrap() {
             russh::ChannelMsg::Data { data } => {
-                assert_eq!(data.to_vec(), [0x00, 0x04, b'P', b'o', b'n', b'g']);
+                assert_eq!(data.to_vec(), b"\x00\x04Pong");
             }
             msg => panic!("Unexpected message {msg:?}"),
         }
@@ -208,13 +205,10 @@ async fn udp_allow_requested_ports() {
         .await
         .expect("Local forwarding failed");
     if timeout(Duration::from_secs(5), async {
-        channel
-            .data(&[0x00, 0x04, b'P', b'i', b'n', b'g'][..])
-            .await
-            .unwrap();
+        channel.data(&b"\x00\x04Ping"[..]).await.unwrap();
         match &mut channel.wait().await.unwrap() {
             russh::ChannelMsg::Data { data } => {
-                assert_eq!(data.to_vec(), [0x00, 0x04, b'P', b'o', b'n', b'g']);
+                assert_eq!(data.to_vec(), b"\x00\x04Pong");
             }
             msg => panic!("Unexpected message {msg:?}"),
         }
@@ -256,15 +250,11 @@ impl russh::client::Handler for SshClient {
         tokio::spawn(async move {
             match &mut channel.wait().await.unwrap() {
                 russh::ChannelMsg::Data { data } => {
-                    assert_eq!(data.to_vec(), [0x00, 0x04, b'P', b'i', b'n', b'g']);
+                    assert_eq!(data.to_vec(), b"\x00\x04Ping");
                 }
                 msg => panic!("Unexpected message {msg:?}"),
             }
-            channel
-                .data(&[0x00, 0x04, b'P', b'o', b'n', b'g'][..])
-                .await
-                .unwrap();
-            channel.eof().await.unwrap();
+            channel.data(&b"\x00\x04Pong"[..]).await.unwrap();
         });
         Ok(())
     }