+++ tests/integration/udp_reject_port_above_max.rs
@@ -1,7 +1,4 @@
-use std::{
-    sync::Arc,
-    time::{Duration, Instant},
-};
+use std::{sync::Arc, time::Duration};
 
 use clap::Parser;
 use russh::keys::{key::PrivateKeyWithHashAlg, load_secret_key};
@@ -11,17 +8,16 @@ use russh::{
 };
 use sandhole::{ApplicationConfig, entrypoint};
 use tokio::{
-    io::AsyncReadExt,
     net::TcpStream,
     time::{sleep, timeout},
 };
 
 use crate::common::SandholeHandle;
 
-/// This test ensures that no more TCP connections from the same IP
-/// than the specified limit are able to connect at the same time.
+/// This test ensures that forwarding is not allowed for UDP ports above the
+/// u16 limit.
 #[test_log::test(tokio::test(flavor = "multi_thread"))]
-async fn tcp_ip_connections_limit() {
+async fn udp_reject_port_above_max() {
     // 1. Initialize Sandhole
     let config = ApplicationConfig::parse_from([
         "sandhole",
@@ -52,15 +48,16 @@ async fn tcp_ip_connections_limit() {
             std::env::var("CARGO_MANIFEST_DIR").unwrap()
         )),
         "--disable-directory-creation",
-        "--listen-address=::",
+        "--listen-address=127.0.0.1",
         "--ssh-port=18022",
         "--http-port=18080",
         "--https-port=18443",
         "--acme-use-staging",
+        "--bind-hostnames=none",
         "--allow-requested-ports",
         "--idle-connection-timeout=1s",
         "--authentication-request-timeout=5s",
-        "--max-simultaneous-connections-per-ip=1",
+        "--http-request-timeout=5s",
     ]);
     let _sandhole_handle = SandholeHandle(tokio::spawn(async move { entrypoint(config).await }));
     if timeout(Duration::from_secs(5), async {
@@ -74,7 +71,7 @@ async fn tcp_ip_connections_limit() {
         panic!("Timeout waiting for Sandhole to start.")
     };
 
-    // 2. Start SSH client that will be proxied
+    // 2. Start SSH client that will fail to bind
     let key = load_secret_key(
         std::path::PathBuf::from(std::env::var("CARGO_MANIFEST_DIR").unwrap())
             .join("tests/data/private_keys/key1"),
@@ -99,46 +96,89 @@ async fn tcp_ip_connections_limit() {
             .success(),
         "authentication didn't succeed"
     );
-    session
-        .tcpip_forward("localhost", 12345)
-        .await
-        .expect("tcpip_forward failed");
+    assert!(
+        session
+            .tcpip_forward("udp.sandhole", u16::MAX.into())
+            .await
+            .is_ok(),
+        "should allow binding ports lesser than or equal to 65535"
+    );
+    assert!(
+        session
+            .tcpip_forward("udp.sandhole", 1u32 + u16::MAX as u32)
+            .await
+            .is_err(),
+        "shouldn't allow binding ports above 65535"
+    );
+    assert!(session.is_closed());
 
-    // 3. Start long-running request that takes the spot for the IP
-    tokio::time::sleep(Duration::from_millis(500)).await;
-    let started = Instant::now();
-    let mut tcp_stream = TcpStream::connect("127.0.0.1:12345")
+    // 3. Start SSH client that will fail to cancel a remote forwarding
+    let key = load_secret_key(
+        std::path::PathBuf::from(std::env::var("CARGO_MANIFEST_DIR").unwrap())
+            .join("tests/data/private_keys/key1"),
+        None,
+    )
+    .expect("Missing file key1");
+    let ssh_client = SshClient;
+    let mut session = russh::client::connect(Default::default(), "127.0.0.1:18022", ssh_client)
         .await
-        .expect("TCP connection failed");
-    let jh = tokio::spawn(async move {
-        let mut data = [0u8; 10];
-        tcp_stream.read_exact(&mut data).await.unwrap();
-        assert_eq!(data, b"0123456789"[..]);
-    });
+        .expect("Failed to connect to SSH server");
+    assert!(
+        session
+            .authenticate_publickey(
+                "user",
+                PrivateKeyWithHashAlg::new(
+                    Arc::new(key),
+                    session.best_supported_rsa_hash().await.unwrap().flatten()
+                )
+            )
+            .await
+            .expect("SSH authentication failed")
+            .success(),
+        "authentication didn't succeed"
+    );
+    assert!(
+        session
+            .cancel_tcpip_forward("udp.sandhole", 1u32 + u16::MAX as u32)
+            .await
+            .is_err(),
+        "shouldn't allow unbinding ports above 65535"
+    );
+    assert!(session.is_closed());
 
-    // 4. Start request that gets rate-limited from IP connection exhaustion
-    tokio::time::sleep(Duration::from_millis(500)).await;
-    let mut tcp_stream = TcpStream::connect("127.0.0.1:12345")
+    // 4. Start SSH clients that will fail to local forward an invalid port
+    let key = load_secret_key(
+        std::path::PathBuf::from(std::env::var("CARGO_MANIFEST_DIR").unwrap())
+            .join("tests/data/private_keys/key1"),
+        None,
+    )
+    .expect("Missing file key1");
+    let ssh_client = SshClient;
+    let mut session = russh::client::connect(Default::default(), "127.0.0.1:18022", ssh_client)
         .await
-        .expect("TCP connection failed");
-    let mut data = [0u8; 10];
-    assert!(tcp_stream.read_exact(&mut data).await.is_err());
-
-    // 5. Start request from different IP that succeeds
-    let mut tcp_stream = TcpStream::connect("[::1]:12345")
-        .await
-        .expect("TCP connection failed");
-    let mut data = [0u8; 10];
-    assert!(started.elapsed() < Duration::from_secs(5));
-    tcp_stream.read_exact(&mut data).await.unwrap();
-    assert!(started.elapsed() > Duration::from_secs(5));
-    assert_eq!(data, b"0123456789"[..]);
-
-    timeout(Duration::from_secs(10), async move {
-        jh.await.unwrap();
-    })
-    .await
-    .expect("timeout waiting for join handle to finish");
+        .expect("Failed to connect to SSH server");
+    assert!(
+        session
+            .authenticate_publickey(
+                "user",
+                PrivateKeyWithHashAlg::new(
+                    Arc::new(key),
+                    session.best_supported_rsa_hash().await.unwrap().flatten()
+                )
+            )
+            .await
+            .expect("SSH authentication failed")
+            .success(),
+        "authentication didn't succeed"
+    );
+    assert!(
+        session
+            .channel_open_direct_tcpip("udp.sandhole", 1u32 + u16::MAX as u32, "127.0.0.1", 12345)
+            .await
+            .is_err(),
+        "shouldn't allow binding ports above 65535"
+    );
+    assert!(session.is_closed());
 }
 
 struct SshClient;
@@ -163,10 +203,7 @@ impl russh::client::Handler for SshClient {
         _session: &mut Session,
     ) -> Result<(), Self::Error> {
         tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_secs(5)).await;
-            channel.data(&b"0123456789"[..]).await.unwrap();
-            channel.eof().await.unwrap();
-            channel.close().await.unwrap();
+            channel.data(&b"\x00\x13Hello, world!"[..]).await.unwrap();
         });
         Ok(())
     }